Fixing Bugs, Added TaskDelay

+++ macros.h
@@ -3,7 +3,7 @@
 #include "RTOS_Config.h"
 #include <stdint.h>
 
-#define MS_TO_TICKS(TimeInMS) (     (uint32_t)(TimeInMS) * (uint32_t)(TARGET_FREQUENCY) / (uint32_t)(1000)      )
+#define MS_TO_TICKS(TimeInMS) (     ( (uint32_t)(TimeInMS) ) * (  (uint32_t)(TARGET_FREQUENCY/1000) )      )
 
 
 #endif
